Increased performance and improved MSVC compatibility
- Ray intersection tests are now cached when traversing K-D trees.
- Now compiles with MSVC (but doesn't run).
- All arithmetic type conversions that can result in data loss, have
  been made explicit, so that MSVC does not produce warnings for them.
  To that end, the type representing dimension has been changed from int
  to size_t, to reduce the number of type conversions needed.

+++ src/py_common.cpp
@@ -228,7 +228,7 @@ PyObject *immutable_copy_func(PyObject *self,PyObject*) noexcept {
     return self;
 }
 
-int setup_buffer(PyObject *self,Py_buffer *view,int flags,void *data,const char *format,int item_size,int length) {
+int setup_buffer(PyObject *self,Py_buffer *view,int flags,void *data,const char *format,size_t item_size,size_t length) {
     view->obj = nullptr;
     if((flags & PyBUF_WRITABLE) == PyBUF_WRITABLE) {
         PyErr_SetString(PyExc_BufferError,"Object is not writable");
@@ -239,11 +239,11 @@ int setup_buffer(PyObject *self,Py_buffer *view,int flags,void *data,const char 
     if((flags & PyBUF_FORMAT) == PyBUF_FORMAT) view->format = const_cast<char*>(format);
 
     view->buf = data;
-    view->len = length * item_size;
+    view->len = static_cast<Py_ssize_t>(length * item_size);
     view->readonly = 1;
-    view->itemsize = item_size;
+    view->itemsize = static_cast<Py_ssize_t>(item_size);
     view->ndim = 1;
-    view->internal = reinterpret_cast<void*>(intptr_t(length));
+    view->internal = reinterpret_cast<void*>(static_cast<intptr_t>(length));
     view->shape = nullptr;
     if((flags & PyBUF_ND) == PyBUF_ND) view->shape = reinterpret_cast<Py_ssize_t*>(&view->internal);
     view->strides = nullptr;
